Add Database kind
